requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
SQLAlchemy==2.0.35
pydantic==2.8.2
